importer/requirements.txt
python-keystoneclient==3.13.0
python-swiftclient==3.4.0

psycopg2==2.6.2
xlrd
pandas
sqlalchemy